requirements.txt
streamlit
pandas
scikit-learn
requests
python-dotenv
gdown